requirements.txt
appdirs==1.4.0
boto==2.45.0
certifi==2017.1.23
dj-database-url==0.4.2
Django==1.10.5
django-bootstrap3==8.1.0
django-storages==1.5.2
django-widget-tweaks==1.4.1
et-xmlfile==1.0.1
gunicorn==19.6.0
jdcal==1.3
lxml==3.7.2
openpyxl==2.4.2
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
python-docx==0.8.6
six==1.10.0
virtualenv==15.1.0
whitenoise==3.3.0
